Create interfaces for "debitar" and "creditar" operations. Implements CreditarUseCase scratch. It's still not dealing with concurrency problems.

+++ app/src/main/kotlin/com/br/rinhadebackend2/crebito/models/TransacaoResponse.kt
@@ -1,6 +1,6 @@
 package com.br.rinhadebackend2.crebito.models
 
 data class TransacaoResponse(
-    val limite: Int,
-    val saldo: Int
+    val limite: Long,
+    val saldo: Long
 )
